Moved import code to the provider
